refactor(auth): move safe-storage helpers to internal module
Move getSafeUserStore, getSafeStateStore and their supporting
helpers out of HerculesAuthProvider into a new safe-storage module
that no barrel re-exports, so they stop leaking into the published
package's public types. The provider imports and uses them; the unit
test imports them directly. No runtime behavior changes.
